修正規化翻譯
